Add files via upload
QDD files are the one needed for the different simulations. The saving folders must be created. 
The LG files calculate the Leggett-Garg factor, both as an ensemble average then a time average. In the feedback case the two calculation do not return the same result due to the loss of ergodicity and the dependence on initial conditions.